update logo & move backend files into their own folder
