+++ src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import { HomePageComponent } from './home-page/home-page.component';
 import { HostLoginPageComponent } from './host-login-page/host-login-page.component';
 import { LoginPageComponent } from './login-page/login-page.component';
-import { PlayerPartyPage } from './player-path/player-party-page';
+import { PlayerPartyPage } from './player-party-page/player-party-page';
 import { PartyLogisticsPageComponent } from './party-logistics-page/party-logistics-page.component';
 import { QueuePageComponent } from './queue-page/queue-page.component';
 import { QrPageComponent } from './qr-page/qr-page.component';
