+++ src/app/app.component.html
@@ -1,36 +1,3 @@
-<nav class="navbar" style="background-image: url('assets/nav-bg.jpg'); ">
-    
-
-        <div class="col-group">
-        </div>
-   
-        <div class="col-group">
-            <a class="navbar-brand" href="#">
-                <img src="assets/gotNextLogo.png" class="gotNextLogo" alt="..." width="82" height="83">
-              </a>
-        </div>
-        
-        <div class="col-group">
-            <form class="d-flex">
-                <div class="card" style="max-width: 400px;">
-                    <div class="row g-0">
-                      <div class="col-md-4">
-                        <img src="assets/gotNextLogo.png" class="img-fluid rounded-start" alt="..." width="82" height="83">
-                      </div>
-                      <div class="col-md-8">
-                        <div class="card-body">
-                          <h5 class="card-title">Username</h5>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </form>
-        </div>
-     
-     
-
-  </nav>
-
-
+<app-navbar></app-navbar>
 
 <router-outlet></router-outlet>
