+++ src/app/app.module.ts
@@ -7,17 +7,18 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HomePageComponent } from './home-page/home-page.component';
-import { UserInfoComponent } from './Username-backend-info/user-info/user-info.component';
+import { UserInfoComponent } from './backend/Username-backend-info/user-info/user-info.component';
 import { LoginPageComponent } from './login-page/login-page.component';
 import { initializeApp,provideFirebaseApp } from '@angular/fire/app';
 import { environment } from '../environments/environment';
 import { provideDatabase,getDatabase } from '@angular/fire/database';
-import { PlayerPartyPage } from './player-path/player-party-page';
+import { PlayerPartyPage } from './player-party-page/player-party-page';
 import { PartyLogisticsPageComponent } from './party-logistics-page/party-logistics-page.component';
 import { QueuePageComponent } from './queue-page/queue-page.component';
 import { HostLoginPageComponent } from './host-login-page/host-login-page.component';
-import { PartyInfoComponent } from './Partyname-backend-info/party-info/party-info.component';
+import { PartyInfoComponent } from './backend/Partyname-backend-info/party-info/party-info.component';
 import { QrPageComponent } from './qr-page/qr-page.component';
+import { NavbarComponent } from './navbar/navbar.component';
 
 @NgModule({
   declarations: [
@@ -31,7 +32,8 @@ import { QrPageComponent } from './qr-page/qr-page.component';
     PartyLogisticsPageComponent,
     HostLoginPageComponent,
     PartyInfoComponent,
-    QrPageComponent
+    QrPageComponent,
+    NavbarComponent
   ],
   imports: [
     BrowserModule,
