+++ src/app/backend/Partyname-backend-info/party-info/party-info-model.ts
