+++ src/app/backend/Partyname-backend-info/party-info/party-info.component.spec.ts
