+++ src/app/navbar/navbar.component.html
@@ -1,0 +1,31 @@
+<nav class="navbar" style="background-image: url('assets/nav-bg.jpg'); ">
+    <div class="col-2">
+    </div>
+
+    <!-- Logo -->
+    <div class="col-8">
+        <div class="d-flex justify-content-center">
+            <a class="navbar-brand" href="#">
+                <img src="assets/textTrans.PNG" class="gotNextLogo" alt="..." width="100" height="40">
+            </a>
+        </div>
+    </div>
+    
+    <!-- Username -->
+    <div class="col-2">
+        <form class="d-flex">
+            <div class="card" style="max-width: 400px;">
+                <div class="row g-0">
+                    <div class="col-md-4">
+                        <img src="assets/gotNextLogoTrans.PNG" class="img-fluid rounded-start" alt="..." width="82" height="83">
+                    </div>
+                    <div class="col-md-8">
+                        <div class="card-body">
+                            <h5 class="card-title">Username</h5>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </form>
+    </div>
+</nav>
