+++ src/app/player-party-page/player-party-page.css
